Fixes issue with Imaginary Numbers
MATLAB uses 1i to represent the imaginary number i. However, this was
used inside of a  for loop that used i as the updated variable. This
lead to some confusion. Upon further research, we were able to figure
out that it was using 1i as the imaginary number irregardless of the for
loop variable. To fix this in our Python port, we used the built in 1j
representation of the imaginary number.