Adds Semi-Generalized Ensemble model classes
The classes allow us to create ensemble models in which the underlying
submodels are identical. Model do not have a fit method yet, but it
should be super simple as it should just wrap the keras model fit
method.

Due to issues with creating deep copies of layers and tensors, we are
currently unable to train different submodels as part of the ensemble.
However, for our use case, this shouldnt matter as a channel's number
should not have an effect on the data seen in the channel.