Adds S3SpectraGenerator Class and Some Refactoring
Refactors existing SpectraGenerator class into an abstract class that is
extended by LocalSpectraGenerator and S3SpectraGenerator. They save the
spectra to disk and S3 respectively. Also added gen_info as an attribute
of the Spectra Generator called `metadata`. This was done so that the
S3SpectraGenerator could access the metadata without having to recreate
it every time. The S3 module uses the metadata to create object keys so
that out bucket is easily navigable and well organized.